Troubleshooting "default-2021.dwl" always being included in Changes
Troubleshooting this issue where "default-2021.dwl" always appears as a changeset even when you make no changes. This involves creating a gitignore file, deleting the contents in the UserSettings folder, committing the delete, and pasting back the  contents in the UserSettings folder so that git no longer tracks the "default-2021.dwl" file.

This step is deleting and committing deletion of the folder contents. Will paste back contents in next checkin.